util: move misplaced functions
